+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Switch from CssAbsoluteFilter to CssRelativeFilter (#822)
 - Add statsd metric collection in hc.lib.s3.get_object()
 - Upgrade to cronsim 2.4
+- Update Signal notification template to include more data
 
 ## v2.8.1 - 2023-04-11